SERVER-3771 -- part 2 -- make mongo shell behave like bash
Remove the all-on-one-line command line editing in the shell and replace it
with something more in the style of other command lines such as bash. Allow
prompts to exceed the width of the terminal screen and allow newlines in
them. Strip all other control characters from the prompt text before
displaying it so that embedded backspaces, tabs and other controls don't
display badly or require extra code to emulate. This prevents users from
using escape sequences in their prompts to perform highlighting, for now
at least. Allow command line text to exceed the available width following
the prompt by wrapping onto other lines the way they bash does it. Make the
newline character display as a newline instead of as a linefeed or DOS OEM
character. Increase the amount of common code between Windows and
non-Windows versions and make them behave the same way. Don't refresh the
Windows command line when the shift key is pressed.


Former-commit-id: 1194acb7fb8d4336f8bb9097f0c847b4b2b3338e