SERVER-4458, SERVER-4460 Improve ctrl-C handling in the shell
Repair the incorrect edit I made in "code cleanup" by moving the space
out from between the quotes in 'if ( code.find( "\n\n\n" ) != string::npos )'
Turn off ctrl-C handling by the Console in Windows so we get the ctrl-C
and can decide what to do with it.  Handle it the same way in Linux and
Windows.  Set the 'gotInterrupted' flag on a ctrl-C so the multi-line
code can exit cleanly on a ctrl-C.  This doesn't resolve SERVER-4460
because it doesn't answer the questions about whether to exit the shell
on a ctrl-C nor about what to do if a server command is running.


Former-commit-id: e5bd79b33f310f4ec2f35a71efd715ef92b9680c