SERVER-3844 make exclusive recursive mutex work properly
We were unlocking every time instead of only when the lock count got back to 0.  This threw an exception with the new Slim RW locks.  We apparently never noticed this was broken with boost shared_mutex because boost's unlock() doesn't throw nor return any error status; it just ignores the problem.


Former-commit-id: 7941b98733fcb480bbfaf15ae2097b3eaa6e081c