SCons: Modify the libdeps module to support two methods for linking libdeps into programs.
This is used to fix linking of programs on Windows that depend on static
initializer execution in object files that don't contain symbols referenced
directly by the program at compile time (i.e., Command registration).

It also unifies the compilation of these sorts of programs on Windows and POSIX
systems, eliminating the use of a MergeLibrary when linking mongod and mongos.

This extension to libdeps could also be used to reimagine MergeLibrary to have
an implementation shared between Posix and Windows systems.  The existing
MergeLibrary system uses relocatable objects on Posix systems, and thus could
behave differently from the one on Windows, with respect to the exeuction of
static initializers.  That change is recommended future work.


Former-commit-id: fe7421e8f00f00cc723bbe0144319f6fac274562