SERVER-4967 allow paste of text with leading tabs
Bring back the test that version 2.0 had for ignoring tabs on
empty lines.  This lets users copy the output JavaScript objects
by SpiderMonkey (which can include leading tabs) and paste it back
into the shell without having command completion get in the way.
This "breaks" command completion on empty lines:  the workaround
is to type a space and then press tab.


Former-commit-id: 71178baeb88780ca6092239e14653ff28c62629a