SERVER-5244 Added temporary diagnostics to 2 tests
The 32-bit Windows BuildBot is showing signs of being out of
memory while mapping a file that has been expanded to try to
get a BSON size error (on purpose).  In successful tests, we
get the BSON size error, on failing tests we get a Windows API
failure instead.  I added db.hostInfo(), db.serverStatus() and
db.stats() to push2.js and pushall.js temporarily to see if we
can learn if it's really an out-of-memory failure.


Former-commit-id: ef111a5b999fdb6871775a57fedf4fab07a10df4