Treat now-deleted tests as passed when using smoke.py to test prior failures.
Since the authoritative list of tests is just the list of files present in the
various test directories, deleting a file means that a test is no longer relevant.
If it used to be in the "failed" state, smoking old failures should treat the file's
absence as a sign that it doesn't matter any more, not as a repeat of the failure.


Former-commit-id: 229931979ff59c9092bc39de9f5e72a35207a74e