SERVER-6303 correct locking for initial sync retrying
With multithreaded replication, the call site for retrying if Cloner misses documents
was changed.  We now much grab a write lock before calling shouldRetry().
I also changed the logging to always log when a retry situation is present.
These situations can result if Cloner misses a document because it was updated, grew,
and moved backward on disk.


Former-commit-id: 9284a9a3db7ed4e94827ad1ba323a08ed559d5f4