Implement and test InitializerDependencyGraph.
An InitializerDependencyGraph is a directed acyclic graph (DAG) of named
initialization operations.  Every node in the graph has a unique name, a
behavior function, and a set of prerequisites.  The graph supports two
inspection functions, one to get the behavior function for a node with a given
name, and the other to produce a vector of node names, ordered in a manner that
does not violate any preprequisite dependneces.

InitializerDependencyGraph is exception-free, and because it is for use very
early in process startup, it does no logging itself.

This patch also introduces a utility macro, MONGO_MAKE_STRING_VECTOR, which is
useful for constructing std::vector<std::string> from a sequence of string
literals.  This patch uses it for testing, but subsequent MONGO_INIT-related
work will rely on it, as well.

Part of work on SERVER-5112.


Former-commit-id: 62d77f23004f6d23f36b6a87a4028451fedaf444