SERVER-7344 add interpreterVersion to shell and buildInfo
Add the ability to get a string for the JavaScript engine so that
scripts can know when they are running under SpiderMonkey versus V8.
The shell gets an 'interpreterVersion()' function that returns a string.
The 'buildInfo' command gets a new field named 'interpreterVersion'.
SpiderMonkey returns "SpiderMonkey 1.7"; V8 returns "V8 3.12.19".

Conflicts:
	src/mongo/db/dbcommands_generic.cpp

Former-commit-id: 286066b379cbadda69c61895f1e295fd433e5479