Reimplement PrivilegeSet.
This new implementation embeds in PrivilegeSet the hierarchical privilege
checking algorithm.  This is necessary in order to allow a connection with
multiple authenticated princiapls to correctly resolve whether or not a
command is authorized, given the case where one principal's authority provides
some of the required privileges, and another's provides the rest.

SERVER-7767


Former-commit-id: 0d8b9c0197b6c07f60efc11dcca1caa5fbd441b8