SERVER-7864 Make copyDB, clone, and cloneCollection work with auth.
copyDB and clone can only be run when auth is enabled if the client is
authenticated as an old-style user. New style users using the extended form
privilege documents cannon run copyDB or clone when auth is enabled.  They can
still run cloneCollection (against the target instance), so long as the source
instance doesn't have auth on.

Conflicts:
	src/mongo/db/cloner.cpp

Former-commit-id: ef7e4d41edbea352e40cc81c5c89f5b1a5d57e1f