SERVER-8200 Don't overlap binDataType with genericRCPtr in Value
This required disabling the short string optimization for BinData (but
not for String) and treat all binary data as large. That shouldn't be
much of an issue in practice since BinData tends to be larger than 12
bytes anyway (eg UUID is 16). It should be possible to reenable the
optimization after 2.4.


Former-commit-id: c17b6310b471765bc4f8df97fb76ed451845f49c