SERVER-8429 Fix tests in jstests/tool.
The affected tests depended on a behavior of mongodump and mongorestore in 2.2
and prior, whereby the tools would attempt authentication via the users'
credentials against several databases, until one matched.  This behavior is
risky, because it could leave the tool authenticating as a user with
insufficient privilege to do its task, and it is inconsistent with other tools.

The tool behavior in 2.4 is to authenticate against the target database if one
is supplied, or admin if none is supplied, unless the user overrides this
behavior by passing --authenticationDatabase=dbname to the tool.


Former-commit-id: 94fb9fdef8e9cb5b02016252a9db6285ed0a7b8a