SERVER-9526 a more elegant solution to printout of BinData in log messages
This commit changes the behaviour of BSONElement::toString() to print BinData in the same way as the other element types. It puts a cap on the lenght of the binary data in the same way as for string. Using full == true overrides the cap in the same way.

Signed-off-by: Greg Studer <greg@10gen.com>
(cherry picked from commit 54845ac08d80b9556cbcbea711ad0529feb09c0f)

Former-commit-id: fb8e784de2a4333166a14dbc0b31d66427a20b7f