SERVER-9621 Disable jstests/dur/closeall.js on 32-bit Windows
This test fails frequently on 32-bit Windows due to remapPrivateView
being unable to remap a file when another thread allocate memory
while a file is unmapped.

Disable the test in 32-bit Windows since this can't really be fixed
without changes to the storage layer (and 32-bit does not benefit
from stress testing).


Former-commit-id: e739d7e4daaf0e65dd1e55c2116bf53744ad88b3