# This is a combination of 2 commits. # The first commit's message is:
Add a test exposing the issue

# This is the 2nd commit message:

Only apply capped rollback items if the ns is still open. We know it
can only be closed/reopend if the same txn did fileops, so it is safe
to assume that if the ns is open, this txn did not do fileops and so
the in-memory stats must be fixed. Otherwise, this txn did do fileops
and thus the ns is closed because of the nsindex rollback items applied
beforehand.


Former-commit-id: 6f59c87275b03fef1a965b827ad548f43ba05c6e