SERVER-8046 Remove some impossible error condition checks
If a node N has a parent, and you are trying to add a sibling to N, then N is already a child,
and therefore N's parent is provably not a leaf. Remove the redundant error checks for this
condition, but leave them as a dassert, at least for now.


Former-commit-id: 27a57dfab40be7b98fab37c4feac82ece0af5dc2