close #461 add rootTransactionId to the result of db.currentOp()
The original implementation (3a788aa) accessed the root transaction
directly rather than caching its id in the TransactionStack.  This allowed
for infoNoauth to race with the commit/abort of the root transaction and
see bad data (segfaults and assertions were common).  Now instead the
TransactionStack always knows the id of its root transaction and this
value can just be read directly.  It's still racy, but the worst thing
that can happen now is a stale read, rather than a crash.


Former-commit-id: d4977e9962e120050c64c0358878042cd187e798