changed storage::Dictionary to always be in a valid state #499
When Dictionary needed to be open()ed after construction, it allowed for
strange races and the proliferation of extra checks in many other places.
We moved open() into the constructor so that if Dictionary is ever fully
constructed, it's also valid.   To do this, we throw a new exception that
must be handled by any creator of a Dictionary that passes
may_create=false.

We also changed Dictionary to be stored in a shared_ptr to allow for
exception-safe closing.

closes #499


Former-commit-id: 941d79f2c1a5d232ecad80ef8a4e3804ae16729c