changed indexes to just ignore dropDups #353
Instead of refusing to build an index if dropDups is present, we can just
ignore the option and try building the index anyway.  This means that if
there actually are duplicates, we'll fail later on and have wasted some
time, but for users that are just trying to import existing data and
indexes, we should be able to safely ignore the option because mongodb
will have already dropped the duplicates.  This is apparently important
for some gridfs implementation.

fixes #353


Former-commit-id: 563a5b4d0f2d0e01718f1d69440192c0779396fc