changed how we link binaries #983
Instead of linking directly with a number of static libraries, we
now traverse the depgraph and merge all libraries we find into a
single large one, and then link with that using -whole-archive.

This ensures that we don't miss any object files that might contain
commands or parameters or whatnot.

This commit only does mongo, mongod, mongos, and tools.  Tests remain
but I am fried for tonight.


Former-commit-id: 4445b625137eac489b5ffb2d9c995051e4954855